test(l1): move blob frame-tx wire tests to the root test crate
