+++ test/tests/common/frame_tx_validation_tests.rs
@@ -11,8 +11,12 @@ use ethrex_common::types::{
     APPROVE_EXECUTION, APPROVE_EXECUTION_AND_PAYMENT, APPROVE_PAYMENT, Block, BlockBody,
     BlockHeader, ChainConfig, EIP4844Transaction, FRAME_SIG_SCHEME_ARBITRARY,
     FRAME_SIG_SCHEME_SECP256K1, FRAME_TX_MAX_VERIFY_GAS, Frame, FrameMode, FrameSignature,
-    FrameTransaction, FrameValidationError, PrefixShape, Transaction, frame_tx_expiry_verifier,
+    FrameTransaction, FrameValidationError, P2PTransaction, PrefixShape, Transaction,
+    WrappedFrameTransaction, frame_tx_expiry_verifier,
 };
+use ethrex_common::types::{BlobsBundle, TxType};
+use ethrex_rlp::decode::RLPDecode;
+use ethrex_rlp::encode::RLPEncode;
 
 /// EIP-4844 `VERSIONED_HASH_VERSION_KZG`. The constant itself lives in a private
 /// module of `ethrex-common`, so it is restated here.
@@ -641,4 +645,102 @@ fn static_validation_requires_the_calldata_floor_to_be_reserved() {
     // Enough frame gas to cover the floor makes it valid again.
     tx.frames[1].gas_limit = 4096;
     assert!(tx.validate_static_constraints().is_ok());
+}
+
+// ---------------------------------------------------------------------------
+// Blob-carrying frame transactions on the wire (EIP-8141 §Networking)
+// ---------------------------------------------------------------------------
+
+/// A blob-carrying frame tx fixture plus a matching sidecar shape. The bundle
+/// contents are not KZG-valid; these tests cover the wire discrimination, which
+/// runs before any cryptographic check.
+fn frame_tx_with_sidecar_shape() -> (FrameTransaction, BlobsBundle) {
+    let mut tx = make_test_frame_tx();
+    let mut hash = [0xABu8; 32];
+    hash[0] = VERSIONED_HASH_VERSION_KZG;
+    tx.blob_versioned_hashes = vec![H256(hash)];
+    tx.max_fee_per_blob_gas = U256::from(7u64);
+    let bundle = BlobsBundle {
+        blobs: vec![],
+        commitments: vec![],
+        proofs: vec![],
+        version: 1,
+    };
+    (tx, bundle)
+}
+
+#[test]
+fn p2p_blob_carrying_frame_transaction_roundtrips_wrapped() {
+    // EIP-8141 §Networking: a frame tx with blobs is wrapped per EIP-7594, as
+    // `[tx_payload_body, wrapper_version, blobs, commitments, cell_proofs]`.
+    let (tx, blobs_bundle) = frame_tx_with_sidecar_shape();
+    let original = P2PTransaction::FrameTransactionWithBlobs(WrappedFrameTransaction {
+        tx,
+        wrapper_version: Some(1),
+        blobs_bundle,
+    });
+
+    let encoded = original.encode_to_vec();
+    let (decoded, rest) = P2PTransaction::decode_unfinished(&encoded).unwrap();
+    assert!(rest.is_empty());
+    assert_eq!(decoded, original);
+    assert_eq!(decoded.tx_type(), TxType::Frame);
+    // `encode_canonical_len` must account for the sidecar, since the wrapped
+    // variant encodes it (the announced pooled-tx size depends on this).
+    assert_eq!(
+        decoded.encode_canonical_len(),
+        decoded.encode_canonical_to_vec().len()
+    );
+
+    // The sidecar is not part of the transaction's identity: the hash matches
+    // the same transaction sent unwrapped.
+    let P2PTransaction::FrameTransactionWithBlobs(ref wrapped) = decoded else {
+        panic!("expected the wrapped variant");
+    };
+    assert_eq!(
+        decoded.compute_hash(),
+        P2PTransaction::FrameTransaction(wrapped.tx.clone()).compute_hash(),
+    );
+
+    // Converting to a plain Transaction would drop the bundle, so it is refused.
+    let as_tx: Result<Transaction, _> = decoded.try_into();
+    assert!(as_tx.is_err());
+}
+
+#[test]
+fn p2p_blob_carrying_frame_transaction_must_be_wrapped() {
+    // Sent unwrapped, a frame tx that declares blobs is rejected: its sidecar
+    // could never be recovered.
+    let (tx, _) = frame_tx_with_sidecar_shape();
+    let encoded = P2PTransaction::FrameTransaction(tx).encode_to_vec();
+    let err = P2PTransaction::decode_unfinished(&encoded).unwrap_err();
+    assert!(
+        format!("{err:?}").contains("must be wrapped with its sidecar"),
+        "unexpected error: {err:?}"
+    );
+}
+
+#[test]
+fn p2p_blobless_frame_transaction_must_not_be_wrapped() {
+    // And the converse: a frame tx with no blobs uses the plain payload, so a
+    // sidecar on the wire is rejected rather than silently ignored.
+    let wrapped = WrappedFrameTransaction {
+        tx: make_test_frame_tx(),
+        wrapper_version: Some(1),
+        blobs_bundle: BlobsBundle {
+            blobs: vec![],
+            commitments: vec![],
+            proofs: vec![],
+            version: 1,
+        },
+    };
+    let mut encoded = vec![TxType::Frame as u8];
+    wrapped.encode(&mut encoded);
+    let mut framed = Vec::new();
+    <[u8] as RLPEncode>::encode(&encoded, &mut framed);
+    let err = P2PTransaction::decode_unfinished(&framed).unwrap_err();
+    assert!(
+        format!("{err:?}").contains("must not carry a sidecar"),
+        "unexpected error: {err:?}"
+    );
 }
